refactored eslintrc
